# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
...
        "/api/order/lists": {
            "get": {
                "tags": [
                    "订单"
                ],
                "summary": "订单列表",
                "operationId": "15d7e4d4253fed1cf19c2e2bed6721df",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "query",
                        "description": "活动id",
                        "required": true
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "订单状态：1 待支付 2支付取消 3支付成功"
                    },
                    {
                        "name": "finished",
                        "in": "query",
                        "description": "是否已成团：1是 ，2否"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/pay/pay": {
            "post": {
                "tags": [
                    "支付"
                ],
                "summary": "支付",
                "operationId": "040427ed825122d43a7c89f2dffafe78",
                "requestBody": {
                    "description": "支付",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "activity_id": {
                                        "description": "活动的id",
                                        "type": "Integer"
                                    },
                                    "type": {
                                        "description": "开团类型必填：1开团 2单独购买",
                                        "type": "Integer"
                                    },
                                    "sign_name": {
                                        "description": "报名学生姓名必填",
                                        "type": "Integer"
                                    },
                                    "sign_mobile": {
                                        "description": "报名手机号必填",
                                        "type": "Integer"
                                    },
                                    "sign_age": {
                                        "description": "报名学生年龄",
                                        "type": "Integer"
                                    },
                                    "sign_sex": {
                                        "description": "性别：1男2女",
                                        "type": "Integer"
                                    },
                                    "course_ids": {
                                        "description": "课程，1，2，3",
                                        "type": "Integer"
                                    },
                                    "is_agree": {
                                        "description": "同意协议 1",
                                        "type": "Integer"
                                    },
                                    "school_child_ids": {
                                        "description": "校区，1，2，3",
                                        "type": "Integer"
                                    },
                                    "info_one": {
                                        "description": "信息一",
                                        "type": "Integer"
                                    },
                                    "info_two": {
                                        "description": "信息二",
                                        "type": "Integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "100000": {
                        "description": "success"
                    }
                }
            }
        },
...
    }
}